Update Ideal Piano.pyw

+++ packages/Ideal Piano.pyw
@@ -1328,33 +1328,37 @@ def midi_file_play(dt):
 
 def initialize(musicsheet, unit_time, start_time):
     global play_midi_file
+    global current_midi_audio
     play_midi_file = False
     playls = []
     start = start_time * unit_time + bars_drop_interval
     if play_as_midi:
         play_midi_file = True
+        if use_soundfont:
+            label.text = 'Rendering current MIDI file with SoundFont, please wait ...'
+            label.draw()
+            window.flip()
         if not if_merge:
             mp.write(musicsheet,
                      60 / (unit_time / 4),
                      start_time=musicsheet.start_time,
                      name='temp.mid')
-            pygame.mixer.music.load('temp.mid')
-            os.remove('temp.mid')
-            os.chdir(abs_path)
+            if use_soundfont:
+                current_waveform = current_sf2.export_midi_file(
+                    'temp.mid', get_audio=True).raw_data
+                current_midi_audio = pygame.mixer.Sound(
+                    buffer=current_waveform)
+            else:
+                pygame.mixer.music.load('temp.mid')
+                os.remove('temp.mid')
+                os.chdir(abs_path)
         else:
             try:
                 if use_soundfont:
-                    label.text = 'Rendering current MIDI file with SoundFont, please wait ...'
-                    label.draw()
-                    window.flip()
                     current_waveform = current_sf2.export_midi_file(
                         path, get_audio=True).raw_data
-                    global current_midi_audio
                     current_midi_audio = pygame.mixer.Sound(
                         buffer=current_waveform)
-                    label.text = ''
-                    label.draw()
-                    window.flip()
                 else:
                     pygame.mixer.music.load(path)
             except:
@@ -1367,6 +1371,10 @@ def initialize(musicsheet, unit_time, start_time):
                         f.write(current_buffer)
                     pygame.mixer.music.load('temp.mid')
                     os.remove('temp.mid')
+        if use_soundfont:
+            label.text = ''
+            label.draw()
+            window.flip()
         pyglet.clock.schedule_once(midi_file_play, bars_drop_interval)
         for i in range(sheetlen):
             currentnote = musicsheet.notes[i]
